# feat: add index on events for faster revision id