--- a/src/migrations/20240208130439-events-revision-id-index.js
+++ b/src/migrations/20240208130439-events-revision-id-index.js
@@ -5,7 +5,7 @@ exports.up = function(db, cb) {
         `
             CREATE INDEX IF NOT EXISTS idx_events_feature_type_id ON events (id)
                 WHERE feature_name IS NOT NULL
-                    AND type IN ('segment-updated', 'feature_import', 'features-imported');
+                    OR type IN ('segment-updated', 'feature_import', 'features-imported');
 
         `,
         cb,