On all pods and instances, we run the same revision update query every second. It is relatively fast when the application has started. This is the single most ran query in unleash.
Benchmarks:

1. Running pod with existing revisionID:
  - old 5.5ms
  - new 0.028ms

2. New pod without existing revisionID
 - old 9.329ms
 - new 0.033ms



This query is getting optimized

https://github.com/Unleash/unleash/blob/7e66a79f9f6f2c6f2475bfd723ea938bbd4fb0f6/src/lib/features/events/event-store.ts#L161